refactore: home 컴포넌트 폴더 추가
