+++ client/src/components/home/index.tsx
@@ -1,0 +1,2 @@
+export { default as Loading } from './Loading';
+export { default as PostsBox } from './PostsBox';
